Split main and GaussianFilter class into two files
